Adding ProductDetailsComponent to declarations
When I was learning Angular, the tutorial of this piece (Adding ProductDetailsComponent to declarations) was missing. As a result, I was stuck with different errors, e.g.
- `Can't bind to 'ngIf' since it isn't a known property of 'div'`
- `Object is possibly 'undefined'.`
- `No pipe found with name 'currency'.`
Which could cause frustration to learners, especially beginners. 

By adding this line, the instruction would be more clear and the learners won't encounter such errors.